requirements.txt
flask==2.3.7
pandas==2.0.3
numpy==1.24.3
gunicorn==21.2.0
Werkzeug==2.3.7
nfl_data_py==0.3.0
pyarrow==12.0.1
python-slugify==8.0.1
streamlit==1.28.1
scipy==1.11.1
scikit-learn==1.3.0
